добавляет действия в reducer, выносит функцию сортировки в utils.js
