+++ project/src/store/reducer.ts
@@ -1,13 +1,15 @@
 import {createReducer} from '@reduxjs/toolkit';
-import {changeCity, fillCityOffers, changeSortOption, setHoveredOfferPin} from './actions';
+import {changeCity, getCityOffers, changeSortOption, setHoveredOfferPin} from './actions';
 import {InitialState} from '../types/state';
 import {offers} from '../mocks/offers';
-import {SortType} from '../const';
+import {SortType, citiesList} from '../const';
 
+const FIRST_CITY = citiesList[0];
 
 const initialState: InitialState = {
-  activeCity: 'Paris',
+  activeCity: FIRST_CITY,
   offers,
+  cityOffers: [],
   sortType: SortType.Popular,
   hoveredOfferPin: null,
 };
@@ -16,9 +18,11 @@ export const reducer = createReducer(initialState, (builder) => {
   builder
     .addCase(changeCity, (state, action) => {
       state.activeCity= action.payload;
+      state.cityOffers = state.offers.filter((offer) => offer.city.name === state.activeCity);
     })
-    .addCase(fillCityOffers, (state, action) => {
+    .addCase(getCityOffers, (state, action) => {
       state.offers = action.payload;
+      state.cityOffers = state.offers.filter((offer) => offer.city.name === state.activeCity);
     })
     .addCase(changeSortOption, (state, action) => {
       state.sortType = action.payload;
