+++ project/src/types/state.ts
@@ -1,6 +1,6 @@
 import {store} from '../store';
 import {Offers, Offer} from '../types/offers';
-import { SortType } from '../const';
+import {SortType} from '../const';
 
 export type State = ReturnType<typeof store.getState>;
 export type AppDispatch = typeof store.dispatch;
@@ -8,6 +8,7 @@ export type AppDispatch = typeof store.dispatch;
 export type InitialState = {
   activeCity: string;
   offers: Offers;
+  cityOffers: Offers;
   sortType: SortType;
   hoveredOfferPin: null | Offer;
 }
